requirements.txt
#fastapi
fastapi
uvicorn

#for environmental variables
python-dotenv

#for email validation
pydantic[email]

#hashing
passlib[bcrypt]

#for creating jwt tokens
python-jose

#for accessing form data
python-multipart

#jinja template
jinja2

#databases
databases
sqlalchemy
asyncpg
psycopg2-binary==2.8.6

#login manager
fastapi-login
